refactor: add addbeans package and move directories
