+++ app/src/main/java/aeropresscipe/divinelink/aeropress/addbeans/ui/AddBeanViewState.kt
@@ -1,4 +1,4 @@
-package aeropresscipe.divinelink.aeropress.beans.ui
+package aeropresscipe.divinelink.aeropress.addbeans.ui
 
 sealed class AddBeanViewState {
     object Initial : AddBeanViewState()
