--- app/src/main/java/aeropresscipe/divinelink/aeropress/beans/BeansTrackerViewModel.kt
@@ -1,8 +1,0 @@
-package aeropresscipe.divinelink.aeropress.beans
-
-import androidx.lifecycle.ViewModel
-import dagger.hilt.android.lifecycle.HiltViewModel
-import javax.inject.Inject
-
-@HiltViewModel
-class BeansTrackerViewModel @Inject constructor() : ViewModel()
