+++ app/src/main/java/aeropresscipe/divinelink/aeropress/beans/ui/BeansTrackerFragment.kt
@@ -1,5 +1,6 @@
-package aeropresscipe.divinelink.aeropress.beans
+package aeropresscipe.divinelink.aeropress.beans.ui
 
+import aeropresscipe.divinelink.aeropress.beans.AddBeanContent
 import aeropresscipe.divinelink.aeropress.ui.theme.AeropressTheme
 import android.os.Bundle
 import android.view.LayoutInflater
